requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# RAG & LLM
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
google-generativeai==0.3.2
openai==1.6.1

# Embeddings & Vector DB
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# Frontend
streamlit==1.29.0
plotly==5.18.0

# Database & Storage
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Data Processing
pandas==2.1.4
